Properly rescale the heavy meson branching ratios to the Scalar.
The branching ratios of SM heavy mesons to the Scalar particle are typically
small, therefore we need to rescale them to make event generation practical.
This is done by applying a uniform scaling such that the largest total
branching ratio of a given heavy meson to the Scalar is equal to unity.

In order to ensure that PYTHIA does not rescale to unity the total branching
ratio of each heavy meson to the Scalar, dummy decay channels to unphysical
states are introduced, such that the uniform scaling is preserved for the final
PYTHIA branching ratios. Placeholder particles of matching charges, with PDG
codes 9900080 and ±9900081, are used to tag these dummy channels.