Add function to configure FairShip for the scalar portal.
`scalar_portal_conf.configure_scalar_portal` is a convenience function
wrapping `FairShipScalarModel` and exposing similar parameters to
`pythia8_conf.configure`. It sets up PYTHIA and ROOT for simulations of the
scalar portal.

Its parameters are the Scalar mass, its mixing angle (coupling) to the SM
Higgs, and the type of heavy mesons from which the Scalar is produced (B or K).

The decay channels of the scalar are inferred from its mass. For now, masses
between 1 and 2 GeV are not supported.

If the `scalar_portal` module is not present in the path, installation
instructions are printed.

Known issues:
* Heavy mesons other than the parent particles of the Scalar are still present,
  and can decay through SM channels.
* PYTHIA seems to be missing several excited kaon states.