Add support for the scalar portal in the `run_simScript.py` macro.
This model can be selected by passing the `--ScalarPortal` switch.

A new option is introduced: `--production-from=`, to choose the type of heavy
mesons used to produce the Scalar. Its possible values are currently 'B' and
'K', with 'B' being the default. It replaces `-A`, which should not be used
with the scalar portal.

Other relevant options are `-m|--mass=` and `-c|--coupling|--couplings`.