Disable SM decay channels instead of replacing parent particles.
Instead of overwriting the parent particles of the Scalar to remove their SM
decay channels, just set `onMode = off`. This is more reliable since it does
not require manually entering the particle properties.

This means that the SM decay modes of the parents are still present in the
PYTHIA Particle Data Table, but they are disabled for the event generation
(their onMode is 0).