Disable SM decay channels for all heavy hadrons in the input file.
This is done to avoid unwanted cascade production, which was already taken into
account when generating the input files.

The channels involved in the actual Scalar production may end up disabled
twice, but this is harmless and more bug-proof.

Finally, I have added instructions on how to enable production from kaons once
cascade data is available for strange hadrons.