Compare observed frequencies of Monte-Carlo events to expected ones.
Run a short (truth level) Monte-Carlo simulation to generate some events, then
check whether the observed production and decay branching ratios are consistent
with the ones used to setup Pythia.

Note that for the decay branching ratios to be consistent, one must account for
channels such as S -> multi-mesons, which are not simulated but still
contribute to the total decay width.

Known issues:
* The Pythia output is quite verbose. Not sure how to disable it.
* Using the same global `ROOT::TDatabasePDG` instance for all tests produces
  annoying side-effects.