Add a placeholder decay to Pythia for the multi-meson channel S -> mesons...
Previously, the multi-meson decay of the Scalar was not simulated. This is
problematic since it causes the total branching ratio of simulated channels to
be less than unity, which leads Pythia to rescale all the branching ratios so
that they sum up to one, effectively enlarging the partial widths of the
remaining decay channels.

Adding an unphysical, placeholder channel to Pythia with the multi-meson
branching ratio prevents this issue. Here we use S -> 9900082 9900082, where
the fake particle 9900082 is appropriately named 'MultiMeson'.