Add the ScalarDecaySelection.yaml file.
This file is used to selectively disable decay channels of the Scalar particle.

It acts as a mask. If a channel is set to True, then the default behaviour is
followed, e.g. 'S -> g g' is enabled above 2 GeV but not below. On the other
hand, if a channel is set to False, then it is unconditionally disabled, and it
will not contribute to the total decay width of the Scalar particle.